# docs/openapi.yaml
openapi: 3.0.0
info:
  title: MiniBlog API
  version: 1.0.0
  description: API REST para gestionar autores y posts de MiniBlog

servers:
  - url: http://localhost:3000/api
    description: Servidor local

paths:
  /authors:
    get:
      summary: Obtener todos los autores
      responses:
        '200':
          description: Lista de autores
    post:
      summary: Crear un autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                email:
                  type: string
                bio:
                  type: string
      responses:
        '201':
          description: Autor creado
        '400':
          description: Campos requeridos faltantes
        '409':
          description: Email ya registrado

  /authors/{id}:
    get:
      summary: Obtener autor por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Autor encontrado
        '404':
          description: Autor no encontrado
    put:
      summary: Actualizar autor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                email:
                  type: string
                bio:
                  type: string
      responses:
        '200':
          description: Autor actualizado
        '404':
          description: Autor no encontrado
    delete:
      summary: Eliminar autor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Autor eliminado
        '404':
          description: Autor no encontrado

  /posts:
    get:
      summary: Obtener todos los posts
      responses:
        '200':
          description: Lista de posts
    post:
      summary: Crear un post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, author_id]
              properties:
                title:
                  type: string
                content:
                  type: string
                author_id:
                  type: integer
                published:
                  type: boolean
      responses:
        '201':
          description: Post creado
        '400':
          description: Campos requeridos faltantes o author_id inválido

  /posts/{id}:
    get:
      summary: Obtener post por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post encontrado
        '404':
          description: Post no encontrado
    put:
      summary: Actualizar post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, author_id]
              properties:
                title:
                  type: string
                content:
                  type: string
                author_id:
                  type: integer
                published:
                  type: boolean
      responses:
        '200':
          description: Post actualizado
        '404':
          description: Post no encontrado
    delete:
      summary: Eliminar post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Post eliminado
        '404':
          description: Post no encontrado

          